Start Palladio Application
This application automates the steps to start Palladio. It’s not
perfect yet, as it produces a dialogue box asking you to choose an app
to open the local host link even though it has already opened that link
in your favorite browser. I’ll see if I can get it to stop doing that.
In the meantime, let me know if it works for you guys.